+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ Added
 - docker-compose file to start a rasa core server together with nlu, an action server, and duckling
 - http server (``rasa_core.run --enable-api``) evaluation endpoint
 - ability to add tracker_store using endpoints.yml
+- ability load custom tracker store modules using the endpoints.yml
 - ability to add an event broker using an endpoint configuration file
 
 Changed